Add files via upload
Unit polytope in three strictly positive dimensions. The green vectors have a one to one correspondence with the black vectors; the green vectors are sparser version of the black vectors. Due to being sparser, their convex hull encompasses more of the volume of the unit simplex, therewith decreasing the solution space which is hosted between the unit simplex and the convex hull of the green set of vectors.